feat: split pages and add attendance views
