+++ apps/web/src/App.tsx
@@ -6,8 +6,16 @@ import SuperAdminLayout from './layouts/SuperAdminLayout';
 import AdminLayout from './layouts/AdminLayout';
 import UserLayout from './layouts/UserLayout';
 import SADash from './pages/superadmin/Dashboard';
+import CompanyList from './pages/superadmin/CompanyList';
+import AddCompany from './pages/superadmin/AddCompany';
+
 import AdminDash from './pages/admin/Dashboard';
+import AddUser from './pages/admin/AddUser';
+import UserList from './pages/admin/UserList';
+import AttendanceList from './pages/admin/AttendanceList';
+
 import UserDash from './pages/user/Dashboard';
+import AttendanceRecords from './pages/user/AttendanceRecords';
 
 export default function App() {
   return (
@@ -25,6 +33,8 @@ export default function App() {
         }
       >
         <Route index element={<SADash />} />
+        <Route path="companies" element={<CompanyList />} />
+        <Route path="companies/add" element={<AddCompany />} />
       </Route>
 
       <Route
@@ -38,6 +48,9 @@ export default function App() {
         }
       >
         <Route index element={<AdminDash />} />
+        <Route path="users/add" element={<AddUser />} />
+        <Route path="users" element={<UserList />} />
+        <Route path="attendances" element={<AttendanceList />} />
       </Route>
 
       <Route
@@ -51,6 +64,7 @@ export default function App() {
         }
       >
         <Route index element={<UserDash />} />
+        <Route path="attendance" element={<AttendanceRecords />} />
       </Route>
 
       <Route path="*" element={<Navigate to="/login" replace />} />
