+++ apps/web/src/layouts/SuperAdminLayout.tsx
@@ -4,19 +4,38 @@ import { clearAuth, getUser } from '../lib/auth';
 export default function SuperAdminLayout() {
   const nav = useNavigate();
   const u = getUser();
+
+  const links = [
+    { to: '/superadmin', label: 'Dashboard' },
+    { to: '/superadmin/companies', label: 'Companies' },
+    { to: '/superadmin/companies/add', label: 'Add Company' }
+  ];
+
   return (
-    <div className="min-h-screen bg-gray-50">
-      <header className="flex items-center justify-between px-6 h-14 bg-black text-white">
-        <div className="font-bold">HRMS Superadmin</div>
-        <div className="flex items-center gap-4">
-          <span className="text-sm">{u?.name}</span>
-          <button onClick={() => { clearAuth(); nav('/login'); }} className="text-sm underline">Logout</button>
+    <div className="min-h-screen flex bg-gray-50">
+      <aside className="w-56 bg-black text-white flex flex-col p-4">
+        <div className="font-bold mb-6">HRMS Superadmin</div>
+        <nav className="flex-1 space-y-2">
+          {links.map(l => (
+            <Link key={l.to} to={l.to} className="block hover:underline">
+              {l.label}
+            </Link>
+          ))}
+        </nav>
+        <div className="pt-4 border-t border-gray-700 text-sm">
+          <div className="mb-2">{u?.name}</div>
+          <button
+            onClick={() => {
+              clearAuth();
+              nav('/login');
+            }}
+            className="underline"
+          >
+            Logout
+          </button>
         </div>
-      </header>
-      <nav className="px-6 py-3 bg-gray-100 border-b">
-        <Link className="mr-4" to="/superadmin">Dashboard</Link>
-      </nav>
-      <main className="p-6">
+      </aside>
+      <main className="flex-1 p-6">
         <Outlet />
       </main>
     </div>
